Changed double string formatting to 4 decimals; Fixed precision bug in Vector.norm() method

+++ src/de/tuberlin/dima/aim3/datatypes/Vector.java
@@ -222,7 +222,7 @@ public class Vector {
     for (double element : elements) {
       sumOfPowers += Math.pow(element, p);
     }
-    return Math.pow(sumOfPowers, 1 / p);
+    return Math.pow(sumOfPowers, (double) 1 / p);
   }
 
   /**
@@ -306,7 +306,7 @@ public class Vector {
   public String toString() {
     String[] valueStrings = new String[elements.size()];
     for (int i = 0; i < elements.size(); i++) {
-      valueStrings[i] = String.format("%.2f", elements.get(i));
+      valueStrings[i] = String.format("%.4f", elements.get(i));
     }
     return "[" + StringUtils.join(", ", valueStrings) + "]";
   }
